refactor: reorganize CLI package structure
- Moved CLI-related code from `internal/cli` to a new `cli` package.
- Removed obsolete CLI files: `cli.go`, `inspect.go`, and `serve.go` to streamline the codebase.
- Updated import paths in `main.go` to reflect the new package structure.
