+++ cli/inspect.go
